[TUBEMQ-309] Add POST support to WebAPI, on both "x-www-form-urlencoded" and "multipart/form-data" body format.
